🛠️: refactor `DatasetInformation services` to `typeScript`
